+++ src/features/cxp/hooks/index.ts
@@ -10,6 +10,7 @@ export * from "./useHistorialFactura";
 export * from "./useEditarFacturaProveedorForm";
 export * from "./useCerrarFacturaSinPago";
 export * from "./useAprobarFacturasLote";
+export * from "./useVerificarSatLote";
 export * from "./useCxpDeepLinks";
 export * from "./useTcDofPorFecha";
 export * from "./useFacturaProveedor";